Creating Final Df
Trimmed down Final.csv to the First 95,000 Rows for Training and Testing.
